server/requirements.txt
fitz==0.0.1.dev2
langchain==0.1.4
langchain-community==0.0.19
llama_index==0.9.40
numpy==1.25.2
pandas==2.0.3
pydantic==2.5.0
python-dotenv==1.0.1
scikit_learn==1.4.0
torch==2.1.2
tqdm==4.66.1
weaviate-client==4.4.2
beautifulsoup4==4.12.3
sentence-transformers==2.3.1
